requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.8.0
aiogram>=3.0.0
langdetect>=1.0.9
